+++ src/Symfony/Component/Yaml/CHANGELOG.md
@@ -4,8 +4,12 @@ CHANGELOG
 3.3.0
 -----
 
- * Deprecated support for implicitly parsing non-string mapping keys as strings. Mapping keys that are no strings will
-   lead to a `ParseException` in Symfony 4.0. Use the `PARSE_KEYS_AS_STRINGS` flag to opt-in for keys to be parsed as
+ * Starting an unquoted string with a question mark followed by a space is
+   deprecated and will throw a `ParseException` in Symfony 4.0.
+
+ * Deprecated support for implicitly parsing non-string mapping keys as strings.
+   Mapping keys that are no strings will lead to a `ParseException` in Symfony
+   4.0. Use the `PARSE_KEYS_AS_STRINGS` flag to opt-in for keys to be parsed as
    strings.
 
    Before: